Migrate Query generation out
